Update login to different location
