+++ src/pages/api/login/index.js
@@ -1,7 +1,6 @@
 /* eslint-disable import/no-extraneous-dependencies */
 import bcrypt from 'bcrypt';
-import { getCookie, setCookie } from 'cookies-next';
-import { isCuid } from '@paralleldrive/cuid2';
+import { setCookie } from 'cookies-next';
 import nacl from 'tweetnacl';
 import { arrayToUtf8, utf8ToArray } from 'enc-utils';
 import * as ecies from 'ecies-25519';
@@ -55,43 +54,4 @@ export default async function handler(req, res) {
     default:
       return res.status(405).send('Method not allowed');
   }
-}
-
-export async function getLogin({ req, res }) {
-  // Check if signature is correct
-  let sessionSign;
-  let sessionData;
-  try {
-    sessionSign = Buffer.from(getCookie('session_sign', { req, res }), 'base64');
-  } catch (err) {
-    return new Error('"session_sign" is not a valid signature');
-  }
-  try {
-    sessionData = JSON.parse(getCookie('session_data', { req, res }));
-  } catch (err) {
-    return new Error('"session_data" is not a valid JSON');
-  }
-
-  const { userId } = sessionData;
-  if (!userId
-    || typeof userId !== 'string'
-    || !isCuid(userId)) return new Error('"session_data".userId is not a valid cuid');
-
-  const date = sessionData.genTime;
-  if (!date || typeof date !== 'number') return new Error('"session_data".genTime is not a valid timestamp');
-
-  const isValid = nacl.sign.open(sessionSign, Buffer.from(process.env.AUTH_PUBLIC, 'base64'));
-  if (!isValid) return new Error('The signature is invalid');
-
-  const user = await prisma.user.findUnique({
-    where: {
-      id: userId,
-    },
-  });
-
-  if (!user) return new Error('This user does not exist');
-
-  if (date < user.lastPwChange.getTime()) return new Error('Password has been changed since login');
-
-  return user;
 }
